feat(webcomponents (actionbar)): new optional prop: shouldDisableButton
Providing true to this prop will disable the action bar button. It is only relevant when
shouldHaveButton prop is also set to true as without that prop there is no button to disable to
begin with. The new prop can, for example, be used to always have a button but only enable it when
the end-user has permissions to use the function under it.